Add util for rendering string keys with HTML content
